+++ app/assets/stylesheets/application.css
@@ -14,12 +14,12 @@
  *= require_self
  */
 
- .my-class {
-    font-size: 16px;
-    color: #333;
-  }
-  
-  #my-element {
-    background-color: #eee;
-    padding: 10px;
-  }
+.my-class {
+  font-size: 16px;
+  color: #333;
+}
+
+#my-element {
+  background-color: #eee;
+  padding: 10px;
+}
