+++ app/javascript/greeting.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { fetchGreeting } from './store';
+import { fetchGreeting } from './greetingSlice';
 
 function Greeting() {
   const dispatch = useDispatch();
